+++ check-syntax.sh
@@ -1,3 +1,0 @@
-#!/usr/bin/env bash
-
-rubocop --format simple --config .rubocop_todo.yml